Classes moved to the lib
